Resolves first error, "Invalid write of size 1".
The strlen function used to allocate space for the name does not include space for the null-terminating char.

The strcpy function copies the entire string over to this allocated space, including the null-terminating char.

We resolve the issue by including "+1" in the space allocation -- allowing space for the null-terminating char.